Use size_t for more hash implementations
These all use hash functions that can return a 64bit value on
64bit platforms, so let's leverage that.

There are some places where we do not do this though, e.g. those
types that use the uint index as hash, such as for IndexedString
and those types that use IndexedString internally.

+++ kdevplatform/language/duchain/indexeddeclaration.h
@@ -44,7 +44,7 @@ public:
 
     inline bool operator!=(const IndexedDeclaration& rhs) const { return !operator==(rhs); }
 
-    inline uint hash() const
+    inline size_t hash() const
     {
         if (isDummy())
             return 0;
