+++ kdevplatform/language/duchain/types/indexedtype.h
@@ -94,7 +94,7 @@ public:
     }
 
     /// Access the type's hash value. \returns the hash value.
-    uint hash() const
+    size_t hash() const
     {
         return m_index >> 1;
     }
