+++ kdevplatform/language/duchain/types/integraltype.h
@@ -85,7 +85,7 @@ public:
      */
     QString toString() const override;
 
-    uint hash() const override;
+    size_t hash() const override;
 
     WhichType whichType() const override;
